assignment1/requirements.txt
cryptography
libnacl
aiohttp
aiohttp_apispec>=3.0.0b1
pyOpenSSL
pyasn1
marshmallow
typing-extensions
packaging
apispec>=6.0.0

pyipv8>=3.2.0
